[moe] clean legacy code
