+++ colossalai/legacy/moe/openmoe/benchmark/benchmark_cai.sh
