+++ colossalai/legacy/moe/openmoe/benchmark/benchmark_cai_dist.sh
