+++ colossalai/legacy/moe/openmoe/benchmark/hostfile.txt
