+++ colossalai/legacy/moe/openmoe/benchmark/utils.py
