+++ colossalai/legacy/moe/openmoe/infer.py
