+++ colossalai/legacy/moe/openmoe/infer.sh
