+++ colossalai/legacy/moe/openmoe/model/convert_openmoe_ckpt.sh
