+++ colossalai/legacy/moe/openmoe/model/openmoe_8b_config.json
