+++ colossalai/legacy/moe/openmoe/model/openmoe_base_config.json
