+++ colossalai/legacy/moe/openmoe/test_ci.sh
