+++ colossalai/legacy/moe/openmoe/train.sh
