--- colossalai/moe/__init__.py
@@ -1,5 +1,0 @@
-from .manager import MOE_MANAGER
-
-__all__ = [
-    "MOE_MANAGER",
-]
